vmwatcher: move checker related code to checker.go
