+++ go/src/koding/vmwatcher/checker.go
@@ -1,6 +1,13 @@
 package main
 
-import "log"
+import (
+	"encoding/json"
+	"koding/db/mongodb/modelhelper"
+	"log"
+	"net/http"
+
+	"labix.org/v2/mgo/bson"
+)
 
 type LimitResponse struct {
 	CanStart     bool    `json:"can_start"`
@@ -25,4 +32,47 @@ func checker(username string) *LimitResponse {
 	}
 
 	return &LimitResponse{CanStart: true}
+}
+
+type ErrorResponse struct {
+	Error string `json:"error"`
+}
+
+func checkerHttp(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+
+	accountId := r.URL.Query().Get("account_id")
+	if accountId == "" {
+		writeError(w, "account_id is required")
+		return
+	}
+
+	yes := bson.IsObjectIdHex(accountId)
+	if !yes {
+		writeError(w, "account_id is not valid")
+		return
+	}
+
+	account, err := modelhelper.GetAccountById(accountId)
+	if err != nil {
+		writeError(w, "account_id is not valid")
+		return
+	}
+
+	response := checker(account.Profile.Nickname)
+
+	js, err := json.Marshal(response)
+	if err != nil {
+		writeError(w, err.Error())
+		return
+	}
+
+	w.Write(js)
+}
+
+func writeError(w http.ResponseWriter, err string) {
+	js, _ := json.Marshal(ErrorResponse{err})
+
+	w.WriteHeader(500)
+	w.Write(js)
 }
